Allow for more parallelism when precompiling modules
Instead of locking the whole precomp store whenever we precompile a
module or re-check dependencies, we now use one lock file per target file.
With this multiple processes may precompile different files and store
them in the same precomp store.

Note: it would be even better to not need those .lock files at all and
instead lock the target files. But that caused weird issues with
multiple processes being able to lock the same file simultaneously. So
I'm leaving this next step for a future change.